requirements.txt
discord.py
requests
python-firebase
